requirements.txt
Django==2.1.3
django-extensions==2.1.3
djangorestframework==3.9.0
isort==4.3.4
mccabe==0.6.1
mysqlclient==1.3.13
pkg-resources==0.0.0
pycodestyle==2.4.0
pyflakes==2.0.0
pytz==2018.7
six==1.11.0
